+++ libsoup/soup-message-headers.c
@@ -925,7 +925,7 @@ sort_ranges (gconstpointer a, gconstpointer b)
 	SoupRange *ra = (SoupRange *)a;
 	SoupRange *rb = (SoupRange *)b;
 
-	return ra->start - rb->start;
+	return (int)(ra->start - rb->start);
 }
 
 /* like soup_message_headers_get_ranges(), except it returns:
